pyproject.toml
[project]
name = "private-credit-stress"
dynamic = ["version"]
requires-python = ">=3.11"

dependencies = [
    "bigdata-client>=2.17.0",
    "bigdata-research-tools>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "nest-asyncio>=1.6.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "slowapi>=0.1.9",
]
